Create do.py
When you run this code, you will first see a splash screen with the title "Welcome to the File Processor". The splash screen contains a label, two buttons (Select Files and Process Files), and a drop-down menu. You can use the drop-down menu to select an option, the selected option will be stored in the selected_option variable. The select_files function allows you to select multiple files, and the process_files function prints "Processing selected files" to the console.